ConcentrationDegreeModel/Model_0330/requirements.txt
imbalanced_learn==0.10.1
imblearn==0.0
matplotlib==3.3.2
mediapipe==0.8.11
numpy==1.23.2
opencv_python==4.7.0.72
pandas==1.4.3
Pillow==9.5.0
pynput==1.7.6
scikit_learn==1.2.2
tensorflow==2.12.0
tensorflow_intel==2.11.0
